app: consolidate makoDesktop bridge typing into lib/desktop after rebase
