+++ app/src/components/RegisterPage.tsx
@@ -24,7 +24,7 @@ import { trackEvent } from "../lib/analytics";
 import {
   supportsDesktopBrowserAuth,
   startDesktopBrowserAuth,
-} from "../utils/desktop";
+} from "../lib/desktop";
 
 interface RegisterPageProps {
   onSwitchToLogin: () => void;
